Revamp Cypress custom commands (#16292)
* Extract UI custom commands

* Extract user related custom commands

* Extract overwrites

* Extract permissions related custom commands

* Extract API custom commands

* Extract helpers custom commands

* Extract database related custom commands

* Rename `helpers` group to `visibility`

* Rename custom command `isInViewport` to `isRenderedWithinViewport`
