+++ frontend/test/__support__/e2e/commands/api/dashboard.js
@@ -1,0 +1,4 @@
+Cypress.Commands.add("createDashboard", name => {
+  cy.log(`Create a dashboard: ${name}`);
+  cy.request("POST", "/api/dashboard", { name });
+});
