+++ frontend/test/__support__/e2e/commands/permissions/sandboxTable.js
@@ -1,0 +1,35 @@
+Cypress.Commands.add(
+  "sandboxTable",
+  ({
+    attribute_remappings = {},
+    card_id = null,
+    group_id = 4,
+    table_id = 2,
+  } = {}) => {
+    // Extract the name of the table, as well as `schema` and `db_id` that we'll need later on for `cy.updatePermissionsSchemas()`
+    cy.request("GET", "/api/table").then(({ body: tables }) => {
+      const { name, schema, db_id } = tables.find(
+        table => table.id === table_id,
+      );
+      const attr = Object.keys(attribute_remappings).join(", "); // Account for the possiblity of passing multiple user attributes
+
+      cy.log(`Sandbox "${name}" table on "${attr}"`);
+      cy.request("POST", "/api/mt/gtap", {
+        attribute_remappings,
+        card_id,
+        group_id,
+        table_id,
+      });
+
+      cy.updatePermissionsSchemas({
+        schemas: {
+          [schema]: {
+            [table_id]: { query: "segmented", read: "all" },
+          },
+        },
+        user_group: group_id,
+        database_id: db_id,
+      });
+    });
+  },
+);
