+++ frontend/test/__support__/e2e/commands/ui/button.js
@@ -1,0 +1,3 @@
+Cypress.Commands.add("button", button_name => {
+  cy.findByRole("button", { name: button_name });
+});
