+++ frontend/test/__support__/e2e/commands/ui/icon.js
@@ -1,0 +1,3 @@
+Cypress.Commands.add("icon", icon_name => {
+  cy.get(`.Icon-${icon_name}`);
+});
